backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
edge-tts
SpeechRecognition
aiofiles
pydub
openai
python-dotenv
httpx
